requirements.txt
polygon-api-client
openai
pandas
numpy
ta-lib-binary
python-dotenv
requests

